chore: Rename deps as server deps
