+++ src/buildWorker.ts
@@ -2,7 +2,7 @@
 /// <reference lib="deno.worker" />
 import { nanoid } from "https://cdn.skypack.dev/nanoid@3.1.30?min";
 import { compileScript } from "../utils/compileScripts.ts";
-import { fs, path } from "../deps.ts";
+import { fs, path } from "../server-deps.ts";
 import { renderPage } from "./renderPage.ts";
 import type { BuildWorkerEvent, Components, ProjectMeta } from "../types.ts";
