+++ src/serve.ts
@@ -2,7 +2,7 @@ import { Server } from "https://deno.land/std@0.118.0/http/server.ts";
 import { cache } from "https://deno.land/x/cache@0.2.13/mod.ts";
 import { relative } from "https://deno.land/std@0.118.0/path/mod.ts";
 import { lookup } from "https://deno.land/x/media_types@v2.11.1/mod.ts";
-import { path as _path } from "../deps.ts";
+import { path as _path } from "../server-deps.ts";
 import { compileScript, compileScripts } from "../utils/compileScripts.ts";
 import { getJson, resolvePaths } from "../utils/fs.ts";
 import { trim } from "../utils/string.ts";
